+++ auswertung/searchResultsByUser/README.md
@@ -1,11 +1,3 @@
 # Twitter Search Result Scraper
 
 mit diesen Tools sollte es möglich sein, die Resultate von Twitter Suchabfragen als JSON zurück zu erhalten.
-
-## Beispiel Aufruf:
-
-```js
-getSearchResults(userName, password, query)
-```
-
-Alle Parameter sind Strings.
